requirements.txt
PyFunctional==0.8.0
netifaces==0.10.5
typing==3.5.2.2
